reducer, actions slice로 합침 store에 slice 적용
